requirements.txt
faster-whisper==1.0.1
google-generativeai==0.4.1
opencv-python-headless==4.9.0.80
ipywidgets==8.1.2
pyyaml==6.0.1